requirements.txt
aiohttp==3.7.4.post0
amqp==5.0.5
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.9.3
billiard==3.6.3.0
black==20.8b1
celery==5.0.5
certifi==2020.6.20
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coverage==5.2
Django==3.0.8
django-appconf==1.0.4
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==2.3.1
django-celery-beat==2.2.0
django-extensions==3.0.2
django-redis==4.11.0
django-select2==7.6.2
django-timezone-field==4.1.1
djangorestframework==3.11.0
dnspython==1.16.0
eventlet==0.30.2
flake8==3.8.3
fluent-logger==0.9.6
freezegun==0.3.15
greenlet==1.0.0
gunicorn==20.0.4
idna==2.10
isort==5.0.5
kombu==5.0.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.4.0
msgpack==0.6.2
multidict==5.1.0
mypy-extensions==0.4.3
packaging==20.4
pathspec==0.8.0
pluggy==0.13.1
prompt-toolkit==3.0.17
psutil==5.7.0
psycopg2-binary==2.8.5
py==1.9.0
pycodestyle==2.6.0
pyflakes==2.2.0
pylint==2.5.3
pylint-django==2.0.15
pylint-plugin-utils==0.6
pyparsing==2.4.7
pytest==5.4.3
pytest-black==0.3.10
pytest-cov==2.10.0
pytest-django==3.8.0
pytest-flake8==1.0.6
pytest-freezegun==0.4.1
pytest-isort==1.1.0
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.3
regex==2020.6.8
requests==2.25.1
sentry-sdk==0.16.0
six==1.15.0
slackclient==2.9.3
soupsieve==2.2
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.12.1
yarl==1.6.3
